Add per-arg parsing flags.
Previously, some requests had special parsing code for arguments which are names
(and which are therefore length-limited in compatibility mode) and for the long
last argument style of .ds.  Now, each argument has a potential set of flags and
the parsing code adjusts itself accordingly.  This allows for multiple name
arguments, such as for .rn.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>