Use decimal.Decimal for real numbers.
Certain tests were failing because of the inexact values produced by Python
(probably due to the move to Python 3.2).  Using decimal.Decimal will prevent
this from happening again.  However, the Decimal class does not by default
produce a decimal point and trailing zero if the value is integral.  We want
one, so append one if needed.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>