Gracefully handle problems with -Txml.
If there were stray text or multiple root elements (usually because of using the
macro package incorrectly) with -Txml, there would be an exception that was not
very helpful to the user.  This commit produces a nice, clean error message
that, while not exceptionally helpful, can give at least a nice shove in the
right direction.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>