Don't swallow empty macros.
This provides a fix to our implementation of copy mode.  Previously, if we had
multiple empty macros, only every other macro would be read.  This is because
our parser looks for a newline as part of the text which ends copy mode, but
we've already consumed that newline.  This code removes the need for the newline
until the first EOL is encountered, while still ensuring that the sequence is
not misinterpreted if it's not the first thing on the line.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>